refactor: relink for netlifay 5
